feat: update ec2 instance user data script
