+++ tools/stacks/deploy-listener.sh
@@ -11,7 +11,7 @@ appName=$1
 region=$2
 stage=$3
 
-STACK_NAME="${appName}-StaticWebsite-${stage}"
+STACK_NAME="${appName}-S3ConfigBucket"
 
 # TODO: adjust stack output key if necessary (it is ...)
 S3_BUCKET=$(aws cloudformation describe-stacks --stack-name $STACK_NAME --region "$region" --query "Stacks[0].Outputs[?OutputKey=='S3BucketName'].OutputValue" --output text)
@@ -21,11 +21,19 @@ if [ -z "$S3_BUCKET" ]; then
     exit 1
 fi
 
-S3_PATH="s3://$S3_BUCKET"
-echo "Syncing website files to $S3_PATH"
+S3_PATH="s3://$S3_BUCKET/ec2/$appName/$stage/listener"
+echo "Syncing Listener to $S3_PATH"
+
+# install only production dependencies on the backend (./backend/src)
+cd backend/src
+npm install --omit=dev
+cd ../../
+
+# zip the backend/src directory into .serverless/listener.zip (ommiting the .git)
+zip -r tools/stacks/backend/.serverless/listener-build.zip backend/src -x "*/.git/*"
 
 # Sync the local directory to the S3 bucket
-aws s3 sync ./frontent/public $S3_PATH
+aws s3 cp tools/stacks/backend/.serverless/listener-build.zip $S3_PATH/listener-build.zip
 
 # Check if the sync command was successful
 if [ $? -eq 0 ]; then
